Refactor: Update AppTheme import paths
The import path for `AppTheme` has been changed from `com.washingtondcsquad.tudee.presentation.design.theme.AppTheme` to `com.washingtondcsquad.tudee.presentation.design.AppTheme` in several composable files. This change standardizes the import path for the application's theme.
